pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "image-retrieval-system"
version = "0.1.0"
description = "Image retrieval system using ViT and FAISS"
requires-python = ">=3.10"
dependencies = [
    "torch>=1.12.0",
    "torchvision>=0.13.0",
    "numpy>=1.21.0",
    "faiss-cpu>=1.7.2",
    "Pillow>=8.0.0",
    "tqdm>=4.60.0",
    "matplotlib>=3.5.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "ipykernel>=6.0.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[project.scripts]
image-ret = "src.index_and_retrieve:main"
